вернур рендерпоинт в мейн + ренейм контейнера
