+++ src/main.js
@@ -1,9 +1,11 @@
-import {RenderPositions, renderElement, renderPoint} from './utils/render.js';
+import {RenderPositions, renderElement, replace} from './utils/render.js';
+import PointView from './view/point-view.js';
+import FormView from './view/form-view.js';
 import MenuView from './view/menu-view.js';
 import NoPointsView from './view/no-points-view.js';
 import FiltersView from './view/filters-view.js';
 import SortView from './view/sort-view.js';
-import ContainerForPointsView from './view/container-for-points-view.js';
+import EventsList from './view/events-list.js';
 import TripInfoView from './view/trip-info-view.js';
 import {generatePoint} from './mock/gen-data.js';
 
@@ -25,11 +27,40 @@ renderElement(filtersContainer, filtersElement, RenderPositions.BEFOREEND);
 const contentSectionElement = document.querySelector('.trip-events');
 
 renderElement(contentSectionElement, new SortView(), RenderPositions.BEFOREEND);
-renderElement(contentSectionElement, new ContainerForPointsView(), RenderPositions.BEFOREEND);
+renderElement(contentSectionElement, new EventsList(), RenderPositions.BEFOREEND);
 
+//отрисовка точки
+const renderPoint = (container, pointData) => {
+  const pointElement = new PointView(pointData);
+  const pointEditForm = new FormView('editForm', pointData);
+
+  const onEscKeyDown = (evt) => {
+    if (evt.key === 'Escape' || evt.key === 'Esc') {
+      evt.preventDefault();
+      replace(container, pointElement, pointEditForm);
+      document.removeEventListener('keydown', onEscKeyDown);
+    }
+  };
+
+  pointElement.setOnPointArrowClick(() => {
+    replace(container, pointEditForm, pointElement);
+    document.addEventListener('keydown', onEscKeyDown);
+  });
+
+  pointEditForm.setOnFormSubmit(() => {
+    replace(container, pointElement, pointEditForm);
+    document.removeEventListener('keydown', onEscKeyDown);
+  });
+
+  pointEditForm.setOnFormArrowClick(() => {
+    replace(container, pointElement, pointEditForm);
+    document.removeEventListener('keydown', onEscKeyDown);
+  });
+
+  renderElement(container, pointElement, RenderPositions.BEFOREEND);
+};
 
 const ulList = contentSectionElement.querySelector('.trip-events__list');
-
 
 if (POINTS_COUNT > 0) {
   renderElement(tripInfoContainer, new TripInfoView(points), RenderPositions.AFTERBEGIN);
