+++ src/view/abstract-view.js
@@ -1,6 +1,5 @@
 import {createElementMarkup} from '../utils/render.js';
 
-
 class AbstractView {
   #element = null;
   _callback = {};
