+++ src/view/events-list.js
@@ -1,15 +1,14 @@
 import AbstractView from './abstract-view.js';
 
-
 const createUlTemplate = () => (
   `<ul class="trip-events__list">
     </ul>`
 );
 
-class ContainerForPointsView extends AbstractView  {
+class EventsList extends AbstractView  {
   get template() {
     return createUlTemplate();
   }
 }
 
-export default ContainerForPointsView;
+export default EventsList;
